add a @groupomania.org mail verification
add 30 mb size authorization for video uplaoded
add post attachement fs.removal for update post route
test  on frontend

+++ Groupomania/backend/controllers/user.js
@@ -7,16 +7,15 @@ const password = require("../models/password");
 
 signup = (req, res, next) => {
   // check email white-space, validity and encrypting
-  email = req.body.email.trim();
-  const emailregex = /^(.*)@(.*)$/;
+  let email = req.body.email.trim();
+  const emailregex = /^(.*)@(groupomania.org)$/;
   const emailIsvalid = emailregex.test(email);
-  if (email && emailIsvalid) {
-    emailCrypted = cryptoJs
-      .HmacSHA256(email, process.env.DCRYPTMAIL)
-      .toString();
-  } else {
-    return res.status(400).json({ message: "A valid email is required" });
-  }
+  if (!emailIsvalid) {
+    return res
+      .status(400)
+      .json({ message: "A valid email from groupomania is required" });
+
+  } 
 
   //check userName white-space and prevent injection
   userName = req.body.userName.trim();
@@ -31,7 +30,7 @@ signup = (req, res, next) => {
     .then((hash) => {
       const user = new User({
         userName: userName,
-        email: emailCrypted,
+        email: email,
         //imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
         password: hash,
       });
@@ -44,10 +43,8 @@ signup = (req, res, next) => {
 };
 
 signin = (req, res, next) => {
-  emailCrypted = cryptoJs
-    .HmacSHA256(req.body.email, process.env.DCRYPTMAIL)
-    .toString();
-  User.findOne({ where: { email: emailCrypted } })
+   
+  User.findOne({ where: { email: req.body.email } })
     .then((user) => {
       if (!user) {
         return res.status(401).json({ message: "Utilisateur non trouvé !" });
@@ -81,14 +78,17 @@ signout = (req, res) => {};
 
 getProfile = (req, res, next) => {
   User.findOne({
-    attributes: ["email", "username"],
-    where: { id: req.token.userId },
+    where: { id: req.params.userId },
   })
     .then((user) => {
       if (!user) {
-        res.status(400).json({ error: "search error" });
+        return res.status(400).json({ error: "search error" });
       }
-      res.status(200).json(user);
+      if (user.id == req.token.userId || req.token.isAdmin) {
+        return res.status(200).json(user);
+      } else {
+        return res.status(401).json({ error: "unauthorized" });
+      }
     })
     .catch((error) => res.status(500).json(error));
 };
@@ -107,28 +107,22 @@ modifyProfile = (req, res, next) => {
         //email checking
         if (req.body.email) {
           email = req.body.email.trim();
-          const emailregex = /^(.*)@(.*)$/;
+          const emailregex = /^(.*)@(groupomania.org)$/;
           const emailIsvalid = emailregex.test(email);
-          if (email && emailIsvalid) {
-            emailCrypted = cryptoJs
-              .HmacSHA256(email, process.env.DCRYPTMAIL)
-              .toString();
-          } else {
+          if (!emailIsvalid) {
             return res
               .status(400)
-              .json({ message: "A valid email is required" });
-          }
-          user.email = emailCrypted;
+              .json({ message: "A valid groupomania email is required" });
+          } 
+          user.email = email;
         }
         //check userName white-space
         if (req.body.userName) {
           userName = req.body.userName.trim();
           if (userName.length > 20) {
-            return res
-              .status(400)
-              .json({
-                message: "UserName must be shorter (max 20 characters)",
-              });
+            return res.status(400).json({
+              message: "UserName must be shorter (max 20 characters)",
+            });
           }
           user.userName = userName;
         }
@@ -136,24 +130,20 @@ modifyProfile = (req, res, next) => {
         //hashing password
         if (req.body.password) {
           passwordChecked = password.validate(req.body.password);
-          console.log("there is a password: " + passwordChecked);
           if (passwordChecked) {
-            passwordHash = bcrypt.hash(req.body.password, 10).then((hash) => {
-              user.password = hash;
-            }).catch((error) => res.status(400).json({ error })) ;
+            passwordHash = bcrypt
+              .hash(req.body.password, 10)
+              .then((hash) => {
+                user.password = hash;
+              })
+              .catch((error) => res.status(400).json({ error }));
           } else {
-            return res
-              .status(400)
-              .json({
-                message:
-                  "Le mot de passe devrait contenir entre 6 et 10 charactères, un symbole et au moins une majuscule, une minuscule et un chiffre.",
-              });
+            return res.status(400).json({
+              message:
+                "Le mot de passe devrait contenir entre 6 et 10 charactères, un symbole et au moins une majuscule, une minuscule et un chiffre.",
+            });
           }
         }
-
-        user.id = req.token.userId;
-        //imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
-
         user
           .save()
           .then(() => res.status(201).json({ message: "user modified!" }))
@@ -164,14 +154,22 @@ modifyProfile = (req, res, next) => {
 };
 
 deleteProfile = (req, res) => {
-  User.destroy({
-    where: { id: req.token.userId },
+  User.findOne({
+    where: { id: req.params.userId },
   })
     .then((user) => {
       if (!user) {
         res.status(400).json({ error: "search error" });
       }
-      res.status(200).json({ message: "user deleted" });
+      if (user.id == req.token.userId || req.token.isAdmin) {
+        User.destroy({
+          where: { id: user.id },
+        })
+          .then(() => res.status(200).json({ message: "User deleted" }))
+          .catch((error) => res.status(400).json({ error }));
+      } else {
+        return res.status(401).json({ error: "unauthorized" });
+      }
     })
     .catch((error) => res.status(500).json(error));
 };
